Correcting a major error:
When the script found a duplicate text, it did not insert the text into the database, as expected, but it nevertheless extracted all the words from the text and inserted them into the database. This would inflate the frequecy count of these tokens, potentially causing some hapax legomena to be regarded as more frequent than they are.